Add files via upload
This file contains Python code for unsupervised learning and cluster analysis of molecules loaded from .sdf files, Tanimoto Distance Matrix (TDM) generation testing and analysis, atomic dictionaries, and various other molecular processing and filtering functions. These Python codes are meant to be used in conjunction with a set of molecular datasets such QM9, or from drug databases. The purpose of this code is to learn molecular classes to prepare specialized groupings designed to be used in generative models to learn new drug molecules that conform to certain small drug molecule "rules" such as the Lipinski's Rule of Five.